Move drag-and-drop logic to context
