+++ src/stores/myspaceContext.tsx
@@ -1,13 +1,17 @@
 import React, { createContext, useContext, useState } from 'react'
-import { WidgetType as AddWidgetType } from '../graphql.g'
+import { DragEndEvent } from '@dnd-kit/core'
+import { arrayMove } from '@dnd-kit/sortable'
+import { WidgetType as AddWidgetType, WidgetReorderInput } from '../graphql.g'
 import { useAnalytics } from 'stores/analyticsContext'
 import { WIDGET_TYPES } from 'constants/index'
 import { MySpace, MySpaceWidget, Collection, Widget } from 'types'
 import { useAddCollectionMutation } from 'operations/portal/mutations/addCollection.g'
 import { useAddWidgetMutation } from 'operations/portal/mutations/addWidget.g'
+import { useEditMySpaceMutation } from 'operations/portal/mutations/editMySpace.g'
 
 export type MySpaceContextType = {
   mySpace: MySpace
+  draggableWidgets: MySpaceWidget[]
   initializeMySpace: (mySpace: MySpace) => void
   updateMySpace: (widget: Widget) => void
   isCollection: (widget: MySpaceWidget) => boolean
@@ -18,10 +22,12 @@ export type MySpaceContextType = {
   addGuardianIdeal: () => void
   addFeaturedShortcuts: () => void
   addNewCollection: () => void
+  handleOnDragEnd?: (event: DragEndEvent) => void
 }
 
 export const MySpaceContext = createContext<MySpaceContextType>({
   mySpace: [],
+  draggableWidgets: [],
   initializeMySpace: /* istanbul ignore next */ () => {
     return
   },
@@ -52,6 +58,9 @@ export const MySpaceContext = createContext<MySpaceContextType>({
   addNewCollection: /* istanbul ignore next */ () => {
     return
   },
+  handleOnDragEnd: /* istanbul ignore next */ () => {
+    return
+  },
 })
 
 export const MySpaceProvider = ({
@@ -60,10 +69,12 @@ export const MySpaceProvider = ({
   children: React.ReactNode
 }) => {
   const [mySpace, setMySpace] = useState<MySpace>([])
+  const [draggableWidgets, setDraggableWidgets] = useState<MySpaceWidget[]>([])
   const { trackEvent } = useAnalytics()
 
   const [handleAddCollection] = useAddCollectionMutation()
   const [handleAddWidget] = useAddWidgetMutation()
+  const [handleEditMySpace] = useEditMySpaceMutation()
 
   function isCollection(widget: MySpaceWidget): widget is Collection {
     return widget.type === WIDGET_TYPES.COLLECTION
@@ -83,6 +94,15 @@ export const MySpaceProvider = ({
 
   const initializeMySpace = (mySpace: MySpace) => {
     setMySpace(mySpace)
+    setDraggableWidgets(
+      mySpace
+        .filter(
+          (w) => w.type !== 'FeaturedShortcuts' && w.type !== 'GuardianIdeal'
+        )
+        .map((c) => {
+          return { id: c._id.toString(), ...c }
+        })
+    )
   }
 
   // Rather than passing in the entire mySpace array, I think we can
@@ -144,8 +164,64 @@ export const MySpaceProvider = ({
     })
   }
 
+  const handleOnDragEnd = async (event: DragEndEvent) => {
+    const { active, over } = event
+
+    // If a draggable item is active, and it is over a droppable area when dropped
+    if (over && active.id !== over.id) {
+      const oldIndex = draggableWidgets.findIndex((w) => w.id === active.id)
+      const newIndex = draggableWidgets.findIndex((w) => w.id === over.id)
+
+      const sortedWidgets = arrayMove(draggableWidgets, oldIndex, newIndex)
+      setDraggableWidgets(sortedWidgets)
+
+      // Prepare widgets for mutation by removing the id field
+      // TO DO: fix this type
+      const updatedMySpace: MySpace = sortedWidgets.map(
+        ({ _id, title, type, bookmarks }) => {
+          const updatedBookmarks = bookmarks?.map(
+            ({ _id, url, label, cmsId, isRemoved }) => ({
+              _id,
+              url,
+              label,
+              cmsId,
+              isRemoved,
+            })
+          )
+          return { _id, title, type, bookmarks: updatedBookmarks }
+        }
+      )
+
+      // If mySpace contains a GuardianIdeal widget, add it to the beginning of the array
+      const guardianIdealWidget = mySpace.find(
+        (w) => w.type === WIDGET_TYPES.GUARDIANIDEAL
+      )
+      if (guardianIdealWidget) {
+        const { _id, title, type } = guardianIdealWidget
+        updatedMySpace.unshift({ _id, title, type })
+      }
+
+      // If mySpace contains a FeaturedShortcuts widget, add it to the beginning of the array
+      const featuredShortcutsWidget = mySpace.find(
+        (w) => w.type === WIDGET_TYPES.FEATUREDSHORTCUTS
+      )
+      if (featuredShortcutsWidget) {
+        const { _id, title, type } = featuredShortcutsWidget
+        updatedMySpace.unshift({ _id, title, type })
+      }
+
+      // Perform mutation to update mySpace
+      handleEditMySpace({
+        variables: {
+          mySpace: updatedMySpace as WidgetReorderInput[],
+        },
+      })
+    }
+  }
+
   const context = {
     mySpace,
+    draggableWidgets,
     initializeMySpace,
     updateMySpace,
     isCollection,
@@ -156,6 +232,7 @@ export const MySpaceProvider = ({
     addGuardianIdeal,
     addFeaturedShortcuts,
     addNewCollection,
+    handleOnDragEnd,
   }
 
   return (
